Updated limit-service README to fix knex usage
refs https://github.com/TryGhost/Utils/pull/190#issuecomment-1124618481

We create the connection like this:
```
// initialize knex db connection for the limit service to use when running query checks
const db = knex({
    client: 'mysql',
    connection: {
        user: 'root',
        password: 'toor',
        host: 'localhost',
        database: 'ghost',
    }
});
```
And pass this as the db option to the limit service.

```
// finish initializing the limits service
limitService.loadLimits({limits, subscription, db, helpLink, errors});
```

But I think the limit service expects a db with a `knex` property that returns the connection. Don't know where this convention started, but in Ghost we create a fake object with a knex property:
https://github.com/TryGhost/Ghost/blob/22e13acd6554c8c053c9cb78b7ba5f24ed4b56f6/core/server/data/db/index.js#L1-L11
```
let connection;

Object.defineProperty(exports, 'knex', {
    enumerable: true,
    configurable: true,
    get: function get() {
        connection = connection || require('./connection');
        return connection;
    }
});
```
Where connection is the instance returned by `knex(configure(config.get('database')));`

Which is passed to the limit service:
```
const db = require('../data/db');

// ...

limitService.loadLimits({
    limits: hostLimits,
    subscription,
    db,
    helpLink,
    errors
});
```
In `core/server/services/limits.js`